+++ unlock-js/src/__tests__/setup/prepare-ganache-for-unlock.js
@@ -24,7 +24,11 @@ const log = (message) => {
 serverIsUp(host, port, 1000 /* every second */, 120 /* up to 2 minutes */)
   .then(async () => {
     // Instantiate the walletService
-    const walletService = new WalletService()
+    const walletService = new WalletService({
+      1984: {
+        provider: providerURL,
+      },
+    })
 
     // We connect to a local node and we expect the node to have unlocked accounts
     // which can be used to send transactions