+++ unlock-js/src/__tests__/walletService.test.js
@@ -1,3 +1,4 @@
+/* eslint jest/no-identical-title: 0 */
 import v4 from '../v4'
 import v6 from '../v6'
 import v7 from '../v7'
@@ -11,7 +12,7 @@ let walletService
 
 describe('WalletService (ethers)', () => {
   beforeEach(() => {
-    walletService = new WalletService()
+    walletService = new WalletService({})
   })
 
   describe('gasAmountConstants', () => {